Wrote AllNumbersGreaterThanUtil.py and tests, and moved RemoveVowels.py and tests to cleancode.
