+++ CodingPractice/PythonAssignments/cleancode/AllNumbersGreaterThanUtil.py
@@ -1,0 +1,3 @@
+
+def all_numbers_greater_than(num, list_of_numbers):
+    return all(n > num for n in list_of_numbers)
