+++ CodingPractice/PythonAssignments/cleancode/RemoveVowels.py
